backend/requirements.txt
# Core ML Stack
tensorflow==2.20.0
keras==3.13.2
numpy==2.3.5
pandas==3.0.0
scikit-learn==1.8.0
joblib==1.5.3
shap==0.50.0
lime==0.2.0.1

# FastAPI Backend
fastapi==0.129.0
uvicorn==0.41.0
pydantic==2.12.5
python-multipart==0.0.22

# Database
pymongo==4.16.0
python-dotenv==1.2.1

# TensorFlow Dependencies (explicit to avoid resolver issues)
h5py==3.15.1
protobuf==6.33.5
grpcio==1.78.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2

# Utilities
requests==2.32.5
tqdm==4.67.3